Fix bug in determining when to model correlations.
Previously I was checking the number of terms in the group formula,
rather than the number of coefs the group formula generated. In cases
where a formula has a single term giving rise to multiple
coefficients (e.g. a categorical column) then correlations were not
modelled, which is not the desired behaviour.